support vgs2play command for Windows
