+++ bin_src/makefile.win
@@ -1,5 +1,6 @@
 all:\
 	..\bin \
+	..\bin\vgs2play.exe \
 	..\bin\vgs2mml.exe \
 	..\bin\vgs2bmp.exe \
 	..\bin\vgs2mkpj.exe \
@@ -45,5 +46,10 @@ T=vgs2wav
 T=vgs2mml
 ..\bin\$(T).exe: $(T).c
 	CL /MT $(T).c
+	copy $(T).exe ..\bin\$(T).exe
+
+T=vgs2play
+..\bin\$(T).exe: $(T).c ../template/vgs2soundw.cpp ../template/vgs2api.c ../template/vgs2tone.c ../template/vgs2.h
+	CL /I../template /MT $(T).c ../template/vgs2soundw.cpp ../template/vgs2api.c ../template/vgs2tone.c dxguid.lib dsound.lib user32.lib
 	copy $(T).exe ..\bin\$(T).exe
