+++ bin_src/vgs2play.c
@@ -1,8 +1,12 @@
 /* VGS BGM player cli */
+#ifdef _WIN32
+#include <Windows.h>
+#else
 #include <pthread.h>
 #include <unistd.h>
 #include <errno.h>
 #include <sys/timeb.h>
+#endif
 #include <stdio.h>
 #include <time.h>
 #include "vgs2.h"
@@ -10,11 +14,6 @@
 int bload_direct(unsigned char n,const char* name);
 void bfree_direct(unsigned char n);
 extern struct _PSG _psg;
-
-/* platform common data */
-static unsigned short ADPAL[256];
-static pthread_mutex_t LCKOBJ=PTHREAD_MUTEX_INITIALIZER;
-static int REQ;
 
 /**
  * get secound from text
@@ -55,7 +54,11 @@ int main(int argc,char* argv[])
     }
     
     /* intialize sound system */
+#ifdef _WIN32
+    if(init_sound(NULL)) {
+#else
     if(init_sound()) {
+#endif
         fprintf(stderr,"Could not initialize the sound system.\n");
         return 2;
     }
@@ -149,38 +152,6 @@ RELOAD:
 
 void putlog(const char* fn,int ln,const char* msg,...)
 {
-    va_list args;
-    time_t t1;
-    struct tm* t2;
-    struct timeb tb;
-    char buf[1024];
-    
-    ftime(&tb);
-    t1=tb.time;
-    t2=localtime(&t1);
-    sprintf(buf,"%04d/%02d/%02d %02d:%02d:%02d.%03d "
-            ,t2->tm_year+1900
-            ,t2->tm_mon+1
-            ,t2->tm_mday
-            ,t2->tm_hour
-            ,t2->tm_min
-            ,t2->tm_sec
-            ,tb.millitm
-            );
-    va_start(args,msg);
-    vsprintf(buf+strlen(buf),msg,args);
-    va_end(args);
-    fprintf(stderr,"%s",buf);
-}
-
-void lock()
-{
-    pthread_mutex_lock(&LCKOBJ);
-}
-
-void unlock()
-{
-    pthread_mutex_unlock(&LCKOBJ);
 }
 
 void make_pallet()
