+++ template/.vgs2mkpj.cfg
@@ -13,6 +13,7 @@ cpy vgs2w.cpp src
 cpy vgs2l.c src
 cpy vgs2x.c src
 cpy vgs2sound.c src
+cpy vgs2soundw.cpp src
 cpy miniz.h src
 cpy maincxx.cpp src
 cpy vgs2i.c src/iPhone
