+++ template/vgs2.h
@@ -162,7 +162,11 @@ void make_pallet();
 int gload(unsigned char n,const char* name);
 int eload(unsigned char n,const char* name);
 int bload(unsigned char n,const char* name);
+#ifdef _WIN32
+int init_sound(HWND);
+#else
 int init_sound();
+#endif
 void term_sound();
 
 /*
