+++ template/vgs2w.cpp
@@ -10,7 +10,6 @@
 #include <io.h>
 #include <d3d9.h>
 #include <d2d1.h>
-#include <dsound.h>
 #include "vgs2.h"
 
 /*
@@ -18,11 +17,6 @@
  * macros
  *----------------------------------------------------------------------------
  */
-#define SND_INIT	0
-#define SND_READY	1
-#define SND_EQ		254
-#define SND_END		255
-
 #define VXSIZE 320
 #define VYSIZE 400
 
@@ -45,13 +39,6 @@ static ID2D1BitmapBrush* _lpD2Brush;
 static D2D1_RECT_F _rtD2RectF;
 static D2D1_RECT_U _rtD2RectU;
 
-static LPDIRECTSOUND8 _lpDS=NULL;
-static LPDIRECTSOUNDBUFFER8 _lpSB=NULL;
-static LPDIRECTSOUNDNOTIFY8 _lpNtfy=NULL;
-static DSBPOSITIONNOTIFY _dspn;
-static uintptr_t _uiSnd=-1;
-static BYTE _SndCTRL=SND_INIT;
-
 static RECT _rect = {0, 0, VXSIZE  , VYSIZE  };
 static const RECT _rectConst = {0, 0, VXSIZE  , VYSIZE  };
 static int need_restore=0;
@@ -62,8 +49,6 @@ static int fs_width, fs_height;
 static BOOL _useD2D=FALSE;
 static LPDIRECT3D9(WINAPI* fpDirect3DCreate9)(UINT);
 static HRESULT(WINAPI* fpD2D1CreateFactory)(D2D1_FACTORY_TYPE,REFIID,D2D1_FACTORY_OPTIONS*,ID2D1Factory**);
-
-static CRITICAL_SECTION _lock;
 
 /*
  *----------------------------------------------------------------------------
@@ -83,18 +68,11 @@ static void term();
 static void gterm();
 static void vtrans(int,int*);
 
-static int ds_init(HWND);
-static void ds_main(void* arg);
-static void ds_stop();
-static int ds_wait(BYTE wctrl);
-
 struct binrec {
 	char name[16];
 	int size;
 	char* data;
 };
-
-static CRITICAL_SECTION _csSnd;
 
 /*
  *----------------------------------------------------------------------------
@@ -114,7 +92,6 @@ extern "C" int __stdcall WinMain(HINSTANCE hIns,HINSTANCE hPIns,LPSTR lpCmd,int 
 	int cn,pn,bn;
 	char path[256];
 
-	InitializeCriticalSection(&_lock);
 	CreateDirectory("DATA",NULL);
 	SetCurrentDirectory("DATA");
 	DeleteFile("LOG.TXT");
@@ -196,8 +173,8 @@ extern "C" int __stdcall WinMain(HINSTANCE hIns,HINSTANCE hPIns,LPSTR lpCmd,int 
 	}
 
 	/* initialize sounds */
-	if(ds_init(hwnd)) {
-		putlog(__FILE__,__LINE__,"ds_init error.");
+	if(init_sound(hwnd)) {
+		putlog(__FILE__,__LINE__,"init_sound error.");
 		term();
 		return FALSE;
 	}
@@ -298,7 +275,6 @@ extern "C" int __stdcall WinMain(HINSTANCE hIns,HINSTANCE hPIns,LPSTR lpCmd,int 
 	vgs2_term();
 
 	term();
-	DeleteCriticalSection(&_lock);
 	putlog(__FILE__,__LINE__,"Exit program.");
 	return TRUE;
 }
@@ -667,210 +643,13 @@ static int d3d_init9(HWND hWnd,const RECT* rect)
 
 /*
  *----------------------------------------------------------------------------
- * initialize DirectSound
- *----------------------------------------------------------------------------
- */
-static int ds_init(HWND hWnd)
-{
-	DSBUFFERDESC desc;
-	LPDIRECTSOUNDBUFFER tmp=NULL;
-	HRESULT res;
-	WAVEFORMATEX wFmt;
-
-	/* initialize DirectSound */
-	res=DirectSoundCreate8(NULL,&_lpDS,NULL);
-	if(FAILED(res)) {
-		putlog(__FILE__,__LINE__,"DirectSoundCreate8 failed: %x",res);
-		return -1;
-	}
-	res=_lpDS->SetCooperativeLevel(hWnd,DSSCL_NORMAL);
-	if(FAILED(res)) {
-		putlog(__FILE__,__LINE__,"SetCooperativeLevel failed: %x",res);
-		return -1;
-	}
-	memset(&wFmt,0,sizeof(wFmt));
-	wFmt.wFormatTag = WAVE_FORMAT_PCM;
-	wFmt.nChannels = SAMPLE_CH;
-	wFmt.nSamplesPerSec = SAMPLE_RATE;
-	wFmt.wBitsPerSample = SAMPLE_BITS;
-	wFmt.nBlockAlign = wFmt.nChannels * wFmt.wBitsPerSample / 8;
-	wFmt.nAvgBytesPerSec = wFmt.nSamplesPerSec * wFmt.nBlockAlign;
-	wFmt.cbSize = 0;
-	memset(&desc,0,sizeof(desc));
-	desc.dwSize=(DWORD)sizeof(desc);
-	desc.dwFlags=DSBCAPS_CTRLPOSITIONNOTIFY;
-	desc.dwBufferBytes=SAMPLE_BUFS;
-	desc.lpwfxFormat=&wFmt;
-	desc.guid3DAlgorithm=GUID_NULL;
-	res=_lpDS->CreateSoundBuffer(&desc,&tmp,NULL);
-	if(FAILED(res)) {
-		putlog(__FILE__,__LINE__,"CreateSoundBuffer failed: %x",res);
-		return -1;
-	}
-	res=tmp->QueryInterface(IID_IDirectSoundBuffer8,(void**)&_lpSB);
-	tmp->Release();
-	if(FAILED(res)) {
-		putlog(__FILE__,__LINE__,"QueryInterface failed: %x",res);
-		return -1;
-	}
-	res=_lpSB->QueryInterface(IID_IDirectSoundNotify,(void**)&_lpNtfy);
-	if(FAILED(res)) {
-		putlog(__FILE__,__LINE__,"QueryInterface failed: %x",res);
-		return -1;
-	}
-	_dspn.dwOffset=SAMPLE_BUFS-1;
-	_dspn.hEventNotify=CreateEvent(NULL,FALSE,FALSE,NULL);
-	if((HANDLE)-1==_dspn.hEventNotify || NULL==_dspn.hEventNotify) {
-		putlog(__FILE__,__LINE__,"CreateEvent failed: %x",GetLastError());
-		return -1;
-	}
-	res=_lpNtfy->SetNotificationPositions(1,&_dspn);
-	if(FAILED(res)) {
-		putlog(__FILE__,__LINE__,"SetNotificationPositions failed: %x",res);
-		return -1;
-	}
-
-	/* start sound controller thread */
-	_uiSnd=_beginthread(ds_main,65536,NULL);
-	if(-1L==_uiSnd) {
-		putlog(__FILE__,__LINE__,"QueryInterface failed: %x",res);
-		return -1;
-	}
-	/* wait for ready */
-	if(ds_wait(SND_READY)) {
-		putlog(__FILE__,__LINE__,"Sound-thread has down.");
-		return -1;
-	}
-
-	return 0;
-}
-
-/*
- *----------------------------------------------------------------------------
- * Sound controller thread
- *----------------------------------------------------------------------------
- */
-static void ds_main(void* arg)
-{
-	HRESULT res;
-	LPVOID lpBuf;
-	DWORD dwSize;
-	char buf[SAMPLE_BUFS];
-
-	putlog(__FILE__,__LINE__,"Sound-thread has started.");
-	_SndCTRL=SND_READY;
-
-	while(1) {
-		while(SND_READY==_SndCTRL) {
-			sndbuf(buf,sizeof(buf));
-			dwSize=SAMPLE_BUFS;
-			while(1) {
-				res=_lpSB->Lock(0
-							,SAMPLE_BUFS
-							,&lpBuf
-							,&dwSize
-							,NULL
-							,NULL
-							,DSBLOCK_FROMWRITECURSOR);
-				if(!FAILED(res)) {
-					break;
-				}
-				Sleep(1);
-			}
-			memcpy(lpBuf,buf,dwSize);
-			res=_lpSB->Unlock(lpBuf,dwSize,NULL,NULL);
-			if(FAILED(res)) {
-				putlog(__FILE__,__LINE__,"Unlock failed: %x",res);
-				goto ENDPROC;
-			}
-			ResetEvent(_dspn.hEventNotify);
-			res=_lpSB->SetCurrentPosition(0);
-			if(FAILED(res)) {
-				putlog(__FILE__,__LINE__,"SetCurrentPosition failed: %x",res);
-				goto ENDPROC;
-			}
-			while(1) {
-				res=_lpSB->Play(0,0,0);
-				if(!FAILED(res)) break;
-				Sleep(1);
-			}
-			WaitForSingleObject(_dspn.hEventNotify,INFINITE);
-		}
-		if(SND_EQ==_SndCTRL) break;
-		_SndCTRL=SND_READY;
-	}
-	putlog(__FILE__,__LINE__,"Sound-thread will now stop.");
-	_SndCTRL=SND_END;
-ENDPROC:
-	return;
-}
-
-/*
- *----------------------------------------------------------------------------
- * wait for stopped of sound controller thread
- *----------------------------------------------------------------------------
- */
-static void ds_stop()
-{
-	if(-1==_uiSnd) {
-		return;
-	}
-	if(ds_wait(SND_READY)) {
-		return;
-	}
-	_SndCTRL=SND_EQ;
-	if(ds_wait(SND_END)) {
-		return;
-	}
-	WaitForSingleObject((HANDLE)_uiSnd,INFINITE);
-}
-
-/*
- *----------------------------------------------------------------------------
- * wait for status modification of sound controller thread
- *----------------------------------------------------------------------------
- */
-static int ds_wait(BYTE wctrl)
-{
-	DWORD ec;
-	while(wctrl!=_SndCTRL) {
-		Sleep(10);
-		if(GetExitCodeThread((HANDLE)_uiSnd,&ec)) {
-			if(STILL_ACTIVE!=ec) {
-				return -1;
-			}
-		} else {
-			return -1;
-		}
-	}
-	return 0;
-}
-
-/*
- *----------------------------------------------------------------------------
  * release DirectX objects
  *----------------------------------------------------------------------------
  */
 static void term()
 {
 	/* release DirectSound */
-	ds_stop();
-	if(_lpNtfy) {
-		_lpNtfy->Release();
-		_lpNtfy=NULL;
-	}
-	if((HANDLE)-1==_dspn.hEventNotify || NULL==_dspn.hEventNotify) {
-		CloseHandle(_dspn.hEventNotify);
-		_dspn.hEventNotify=NULL;
-	}
-	if(_lpSB) {
-		_lpSB->Release();
-		_lpSB=NULL;
-	}
-	if(_lpDS) {
-		_lpDS->Release();
-		_lpDS=NULL;
-	}
+	term_sound();
 
 	/* release DirectGraphic */
 	gterm();
@@ -1030,26 +809,6 @@ int loadrom(const char* fname)
 
 	fclose(fp);
 	return 0;
-}
-
-/*
- *----------------------------------------------------------------------------
- * inter thread lock
- *----------------------------------------------------------------------------
- */
-void lock()
-{
-	EnterCriticalSection(&_lock);
-}
-
-/*
- *----------------------------------------------------------------------------
- * inter thread unlock
- *----------------------------------------------------------------------------
- */
-void unlock()
-{
-	LeaveCriticalSection(&_lock);
 }
 
 /*
